+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.10.2
+
+- Fixed nested multiple expands with shared path ([#586](https://github.com/pocketbase/pocketbase/issues/586#issuecomment-1357784227)).
+  A new helper method `models.Record.MergeExpands(map[string]any)` was also added to simplify the expand handling and unit testing.
+
+
 ## v0.10.1
 
 - Fixed nested transactions deadlock when authenticating with OAuth2 ([#1291](https://github.com/pocketbase/pocketbase/issues/1291)).